+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `datacontract import odata` creates a datacontract from OData 4 metadata at an URL or from a local file.
+
 ## [1.2.1] - 2026-09-22
 
 ### Highlights